Added Field Repo/Service/Controller

+++ Data/Repository/FieldRepository.cs
@@ -2,6 +2,7 @@
 using FieldBooking.Data.Models;
 using FieldBooking.Domain.Models;
 using FieldBooking.Domain.Repository;
+using Microsoft.EntityFrameworkCore;
 
 namespace FieldBooking.Data.Repository
 {
@@ -17,27 +18,50 @@ namespace FieldBooking.Data.Repository
 
         }
 
-        public void AddField(FieldDto fieldDto)
+        public async Task<FieldDto> CreateAsync(FieldDto fieldDto)
         {
             var field = _mapper.Map<Field>(fieldDto);
-            _context.Fields.Add(field);
-            _context.SaveChanges();
+            await _context.Fields.AddAsync(field);
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e.Message);
+            }
+
+            return _mapper.Map<FieldDto>(fieldDto);
         }
 
-        public FieldDto GetField(int id)
+        public async Task<FieldDto> GetAsync(int id)
         {
-            return _mapper.Map<FieldDto>(_context.Fields.FirstOrDefault(x => x.Id == id));
+            return _mapper.Map<FieldDto>(await _context.Fields.FirstOrDefaultAsync(x => x.Id == id));
         }
 
-        public List<FieldDto> GetAllField()
+        public async Task<List<FieldDto>> GetAllAsync ()
         {
-            return _mapper.Map<List<FieldDto>>(_context.Fields.ToList());
+            return _mapper.Map<List<FieldDto>>(await _context.Fields
+                .Include(field=>field.Address)
+                .Include(field=>field.Calendar).ToListAsync());
         }
 
-        public void RemoveField(int id)
+        public async Task<FieldDto> RemoveAsync(int id)
         {
-            var field = _context.Fields.FirstOrDefault(x => x.Id == id);
+            var field = await _context.Fields.FirstOrDefaultAsync(field => field.Id == id);
+            if (field is null) throw new ArgumentException($"Can't find user with specified id: {id}");
             _context.Fields.Remove(field);
+            await _context.SaveChangesAsync();
+            return _mapper.Map<FieldDto>(field);
+        }
+
+        public async Task<FieldDto> UpdateAsync(FieldDto fieldDto)
+        {
+            var field = await _context.Fields.FirstOrDefaultAsync(field => field.Id == fieldDto.Id);
+            if (field is null) throw new ArgumentException($"Can't find field in database");
+            _mapper.Map(fieldDto, field);
+            await _context.SaveChangesAsync();
+            return _mapper.Map<FieldDto>(field);
         }
     }
 }
